Make Flow Go SDK model use explicit
