+++ api/rosetta/submit.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/labstack/echo/v4"
 
-	"github.com/onflow/flow-go-sdk"
+	sdk "github.com/onflow/flow-go-sdk"
 
 	"github.com/optakt/flow-dps/rosetta/identifier"
 )
@@ -60,7 +60,7 @@ func (c *Construction) Submit(ctx echo.Context) error {
 		return echo.NewHTTPError(http.StatusBadRequest, invalidFormat(txBodyEmpty))
 	}
 
-	var tx flow.Transaction
+	var tx sdk.Transaction
 	err = json.Unmarshal([]byte(req.SignedTransaction), &tx)
 	if err != nil {
 		return echo.NewHTTPError(http.StatusBadRequest, invalidFormat(txBodyInvalid, withError(err)))
