+++ api/rosetta/transactor.go
@@ -15,7 +15,7 @@
 package rosetta
 
 import (
-	"github.com/onflow/flow-go-sdk"
+	sdk "github.com/onflow/flow-go-sdk"
 
 	"github.com/optakt/flow-dps/rosetta/identifier"
 	"github.com/optakt/flow-dps/rosetta/object"
@@ -25,7 +25,7 @@ import (
 // Parser is used by the Rosetta Construction API to handle transaction related operations.
 type Parser interface {
 	DeriveIntent(operations []object.Operation) (intent *transactor.Intent, err error)
-	CompileTransaction(intent *transactor.Intent, metadata object.Metadata) (tx *flow.Transaction, err error)
-	ParseTransaction(tx *flow.Transaction) (operations []object.Operation, signers []identifier.Account, err error)
-	AttachSignature(unsignedTx *flow.Transaction, signature object.Signature) (tx *flow.Transaction, err error)
+	CompileTransaction(intent *transactor.Intent, metadata object.Metadata) (tx *sdk.Transaction, err error)
+	ParseTransaction(tx *sdk.Transaction) (operations []object.Operation, signers []identifier.Account, err error)
+	AttachSignature(unsignedTx *sdk.Transaction, signature object.Signature) (tx *sdk.Transaction, err error)
 }
